pyproject.toml
[project]
name = "aixis-platform"
version = "2.0.0"
description = "Aixis AI監査プラットフォーム - 日本初の独立系AI検証SaaS基盤"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core audit engine
    "anthropic>=0.42",
    "playwright>=1.49",
    "pydantic[email]>=2.10",
    "pydantic-settings>=2.6",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "numpy>=2.1",
    "rich>=13.9",
    # Web platform
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.12",
    "httpx>=0.28",
    # htmx is loaded via CDN in templates, not a Python package
    # Background tasks
    "arq>=0.26",
    "redis>=5.2",
    # Reports
    "weasyprint>=63.0",
    # CLI
    "typer>=0.15",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.34",
    "ruff>=0.8",
]

[project.scripts]
aixis = "aixis_agent.cli:app"
aixis-web = "aixis_web.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aixis_agent", "src/aixis_web"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
